requirements.txt
streamlit>=1.36
pandas>=2.2
numpy>=1.26
plotly>=5.22
python-dateutil>=2.9
openpyxl>=3.1
python-pptx>=0.6.23